refactor(FON-468): organize pages by entity, keep thin espace shells
