+++ apps/client/src/pages/agenda/AgendaPreviewPage.tsx
