+++ apps/client/src/pages/observations/ObservationDetailsPage.tsx
