+++ apps/client/src/pages/sessions/LolfiRedirectMagistratPage.tsx
