+++ apps/client/src/pages/sessions/ManageSessionPage.tsx
